backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai==1.65.0
boto3
requests==2.31.0
firebase-admin
runpod
